[INTERNAL] sap.ui.test: Migrate gherkin tests to qunit2
Problem:
The Gherking testsuite contained code that complied with
qunit1 only and not with qunit2. In particular:
1) the self-test qunit/gherkin/testHarnessTests.js checked
 for the qunit test output using a selector that is valid
 for qunit1 but not for qunit2
2) the test file gherkin/GherkinWithQUnit/Steps.js from
 the demokit "GherkinWithQUnit" sample used API that
 is available with qunit1 and not with qunit2 (API for
 accessing the qunit "assert" object)

Solution:
Made the above test code agnostic to the qunit version
it ises:
1) for the self-test, switched to a selector that works
both with qunit1 and qunit2
2) for the demokit "GherkinWithQUnit" sample, switched
to API that is available in both qunit and and qunit2
(i.e. now the test accesses the 'assert' object from
the "this" object, as recommended in the documentation)

Also switched to qunit2 in all tests of the Gherkin
suite.

Change-Id: I1e6651f07d684aa4fcbad5294402b91074634da2