[FIX] MDCTable: Reset via p13n/settings dialog
Just like with the SmartTable, the expectation is that the Reset button
in the p13n settings dialog resets the entire state, including column
widths, for example. It should not only reset the enabled p13n options
that are represented by a panel in the dialog, which is the default
behavior of the Engine.
Reset in the filter dialog still only resets the filters.

Also fixes a bug in the column that was detected in the tests for the
personalization util. An error was thrown if the enableColumnResize
property of the table was changed.

Change-Id: Ib247f9c9d2eb58c3b0c9715910c7278704d4a34b
SNOW: DINC0059545