[FIX] sap.f.DynamicPageTitle: Fix overlaping of snappedContent
Problem: When there is a Text control with long text in snappedContent
aggregation of DynamicPageTitle and the screen size hits the tablet
break point, the text gets overlapped by the expand/collapsed buttons.
This happens on all themes.

Solution: There is actually no VD spec for snapped content or at least I
cannot find it. It seems if Label is in the snappedContent - it works
well. It has overflow: hidden style, which makes the snappedContent div
of the DynamicPageTitle somehow bigger. Now same style is applied also
for Text control, if placed there.

SNOW: DINC0088126
Change-Id: If92099eb27a456757ec02099d7244ce669692ecd