[INTERNAL] sap.ui.fl: DelegateMediator refactored delegate registration
Delegate registration is now possible for a specific control. This
allows to register a 'write' delegate for a specific control
additionally to the model specific default 'read' delegate definition.
The read and write delegate registration and getter functions are
all splitted into separate functions.

This is the 3rd openUI5 commit for the delegate mediator refactoring.
The legacy delegate implementation is removed.

Change-Id: I60398c1b664d2f36ba0f20ee0c8ed65683c6510e
SNOW: TEAM42-325