[INTERNAL] ui.table: Action column must be in the scroll viewport
 - The min-width calculation of the Table converts rem width property
values of columns to pixel values and this might cause a rounding
problem. At the end, for large number of columns and in different zoom
level there might be some pixel differences. This change takes 2px width
as an epsilon value to determine whether the dummy column is visible or
not.

SNOW: CS20240007305029

Change-Id: I52ba00c10828680c75c08ef106973b02d9dab969