[INTERNAL] core.Lib: fix deprecation text for DataType pseudo modules
The deprecation message that is logged to the console when a pseudo
module is accessed, should distinguish between DataType instances and
enums. For enums, importing them from the library.js module is the
correct approach. for DataType instances, they should be accessed
via the DataType.getType() method. Additionally, the library.js module
must be imported in that case to ensure that the DataType instance has
been created yet.

The loader now accepts a function instead of a deprecation message as
the correct message can only be determined late.

Additionally, the probing require now also logs the message.

Change-Id: I40a43c4e7656fd2596a6a2e93ba3211f5e6e4016