[INTERNAL] DemoKit: Opa test reworked
Problem:
To navigate to a section ot subsection, the OPA test relies
on a click over a private control inside the ObjectPage.
The test will this break when the type of private control
is changed (as planned).

Solution:
Reworked the OPA test use the ObjectPage public API for
navigation to sections or subsections.

JIRA: BGSOFUIPIRIN-5360
Change-Id: Ide93c4f01cf81ad5caec13aa5ea60c3306d20f26