[FIX] sap.f.FlexibleColumnLayout: Fix wrongly sized mid column
Problem: When switching too fast (for me only when sync switching)
between TwoColumnsMidExpanded and ThreeColumnsMidExpanded layouts of
FCL, the mid column is not resized properly at the end of resizing, so
it does not auto expand, but either have a fixed (not appropriate
width).

Solution: In the afterResize callback, if coulmn is autoSized=true, we
delete the fixed width set (if it, for some reason, hasn't been
deleted).

SNOW: DINC0110982
Change-Id: I297e3a2844742d21a4f9b7ae7f35cadba7c8147a