[INTERNAL] CustomStyleClassSupport (test): fix test instability
The test "add and remove a scope class" still regularly fails.
It might be that Parameters._getScope does not return scopes if the
theme for sap.ui.testlib wasn't applied yet.

By waiting for the theme to be applied and by retrieving a parameter
from the testlib, not from sap.ui.core, this hopefully can be fixed.

Additionally, the test no longer uses sap_hcb as this theme will be
removed with the next major version of UI5. The new 'test_theme_scoped'
is a 1:1 copy of sap_hcb.

Change-Id: I5471e7b0bb4e304669d7c26393efffbd3a32297f