[INTERNAL] sap.ui.test.autowaiter._navigationContainerWaiter: Wait for resize notifications
Problem:
The OPA test [with "autowait" enabled] starts to interact with
the controls on a newly opened page before the ResizeHandler managed to notify the controls [of the newly opened page] of their new size.

Solution:
-The ResizeHandler notifications come from a background progess,
this is why the autowaiter does not wait for them to complete
(i.e. a background process never completes).
-However, the autowaiter *can subscribe to be notified* when the next resize notification is to triggered (via the IntervalTrigger, as its default instance also internally triggers the ResizeHandler).
-Fixed by subscribing the _navigationContainerWaiter for the same IntervalTrigger event that also synchronously triggers the resize notifications (_navigationContainerWaiter is chosen as the issue is observed to happen upon navigation to a new sap.m.NavContainer page).

Change-Id: If0cdab50686e7800c40d27c7c8e3a43acf808e13
SNOW: DINC0111284