[INTERNAL][FIX] sap.ui.fl: Cleanup dependency map when VariantModel is destroyed
During initialization, the VariantModel adds the control changes of all
current variants to the dependency map.
When the VariantModel is destroyed during a soft reload, the dependency
map should be cleaned up to make sure that no incorrect changes are left
in the map when the app loads again.
This is necessary since the current variant might have switched which
would result in an inconsistent UI state where dependencies exist that
don't belong to the current variant.

SNOW: DINC0115286
Change-Id: I922155c491211a1606598a8f5dd7eace0dc00c4a