[FIX] sap.m.TimePickerClock: fixed time selection when the page is scrolled
Problem: until now, when the page is not scrolled, the clock number
selection with mouse worked, but when the page is scrolled, the
selection didn't work as expected because scroll position wasn't been
respected.

Solution: now the scrollTop/scrollLeft positions are used in the
calculations and the clock number selection works always.

SNOW: DINC0101661
Change-Id: Ibedbb89b7d29e896a6d9b7ab2c2f8656bc7dfc5a