[INTERNAL] sap.ui.rta: introduce depth for fetching bindings for reveal
When revealing a control, the bindings of the control and children are fetched to
check if the control would be hidden via annotations. For Container controls
only a certain depth of children have an influence over the visibility of
the container. This change introduces a depth parameter in the designtime
for the reveal action.
This new parameter is implemented for the ObjectPageSection and ObjectPageSubSection.

SNOW: DINC0080529
Change-Id: I8c8fe37b535845d57ee281161810c1cfbb40f940