[FIX] sap.m.OverflowToolbar: Fix closing Popover with ComboBox
Problem: When ComboBox with layoutData closeOnInteraction=false is
placed in the Popover of the OverflowToolbar, the Popover is closed upon
iteraction with the ComboBox, while it should not be closed (as per the
layoutData setting). An error in the console is also logged.

Solution: It seems, some time ago, the ComboBox as introduces new
private property "_open", which replaces the old one (now deprecated)
"open" property. When this property is set, the OverflowToolbar should
not be invalidated, as this property has no effect on the width of the
ComboBox. Also new additional property - effectiveSHowClearIcon, is
added in the noInvalidationProps of OverflowToolbar config, as it does
not affect the width of the control (clear icon is shown inside the
ComboBox, next to the open arrow icon).

SNOW: DINC0116022
Change-Id: Ie8827e4876b6868e2eb86d174295c6f03368c592