[INTERNAL] sap.ui.core: Add transformation hints where applicable
Currently, the only implemented hint is replace-param which can replace
a method or function parameter with a constant boolean value. The
parameter itself is removed if it is the last parameter in the signature
or renamed (prefixed with an underscore) otherwise.

If multiple parameters are to be removed, they should be listed in
reverse order so that the hint for the last parameter in a signature is
replaced (and removed) first.

The order of parameters has been changed for some local functions in XMLTemplateProcessor, in order to allow to replace bAsync more easily.

JIRA: CPOUI5FRAMEWORK-618
Change-Id: I98838a8b5080fc721a87b81a4abd60aa16340aa0