[FIX] sap.m.DynamicDateRange: fix footer buttons visibility when there is not enough space
Problem: There was a regression called by change #6109759 which led
to invisible footer buttons of the DDR when there is not enough
screen height for the whole value help popover. The issue was because
of removed invalidate() method call of the popover on change of the
footer buttons visibility.

Solution: Now invalidate() is called when footer buttons visibility
is set to true, which solves the issue. Calling of invalidate() when
footer buttons visibility is set to false is not needed as this
happens only on popover close.

Snippix for test: #471899

SNOW: DINC0114733
Change-Id: Ic38d6c7bc87d7824681efa7423c3d4db82530aaa