[FIX] sap.m.Avatar: Fix image loading of Avatar
Problem: When the Avatar first receives an empty string or an invalid
image url for its "src" property, and then synchronously - a valid one,
the valid image is not shown.

Solution: The problem happens because of a timing - as the invalid and
valid images are requested at the same time (synchronously), it may
take more time for the error callback to be called (when no image is
found). When the error callback is called after the onLoad callback,
the CSS class sapFAvatarImage is removed and the image is not visible.
Now in the error callback we check if the current "src" of Avatar is
still the same as the one that caused the error callback call. If not
the same, we do not clean the CSS --> obviously a new "src" has been set
in the meantime.
Additionally, no need to validate the "src" property, when it is an
empty string.

SNOW: DINC0119517
Change-Id: Ib19c087ffa827d324f2b296e9640aacf3600f92b