[INTERNAL][FIX] sap.ui.rta: Avoid unnecessary checks when detaching the RenameHandler
There are many scenarios that trigger a selection change which requires
to detach old RenameHandlers and attach new ones. In some cases, the
selection change is triggered because the element of the previously
attached overlay was destroyed, which will break additional checks In
_manageClickEvent.
Since we only need these additional checks when attaching new handlers,
avoid them and just detach directly.

SNOW: DINC0083211
Change-Id: Ia9eef8e88231f9efbbe8301878bf9d519c7638da