[FIX] sap.m.OverflowToolbar: Aria-roledescription is removed
The aria-roledescription has been removed from the sap.m.OverflowToolbar.
Previously, the announcement of 'Overflow' in the toolbar seemed to be redundant
for the screen reader users, since they are not aware about the items that are hidden
in the overflow area.
Now, only the role is announced for the sap.m.OverflowToolbar.

SNOW: DINC0111000
Change-Id: I2127be2d41d623d3086aca2c88b04e5441c34396