[FIX] sap.m.Toolbar: Fix left aligment in Tables
Problem: Toolbar with different than Transparent Design, placed in
Tables, does not have correct left aligment (1 rem) in all three themes
- Belize, Quartz and Horizon.

Solution: It seems, in the past Transparent Design has been used for
Toolbars in Tables and the specific styles for Toolbar in Table, were
applied to the Transparent Toolbar. However, now people are creating
Toolbars with different designs in Tables, so that we need to style all
of them, regardless of the design, according to the Toolbar with Title
specification. In all three themes the left-padding should be 1 rem.

SNOW: DINC0110919
Change-Id: If634ac4515817e5732057245d3261549eb0aff07