[FIX] FCL: Restore focus same column fixed
When the focus is previously in the column, that stays current,
after layout change/columns resize the focusout event never
fires and the focus for the column doesn't get updated,
so FCL always restore the focus to the first focusable element.
If we first check where the focus is we can omit the redundant focusing.

BCP: 2370131414
Change-Id: Iab6326c841c1c83c47c6dc09495083611a6f44c8