[INTERNAL] dnd: keyboard support for aggregation reordering
 - The keyboard combination Ctrl+Arrow Left/Right and Ctrl+Arrow Up/Down
should allow aggregation reordering if drag-and-drop is enabled for
aggregation reordering.
 - Central DragAndDrop handler detects whether aggregation reordering is
enabled via Drag/DropInfo configuration and mimics the dragstart,
dragenter, drop, dragend events with custom events.
 - Those custom events are marked with "simulated" keys. This
information can be used by control and default can be prevented in
dragstart event and in this case other drag events are not dispatched.
 - This event simulation was necessary since complete logic of
drag-and-drop relay on those events. This is not only true for the
central drag-and-drop handler but also for other drag-and-drop consumers
e.g. TreeTable does not allow column reordering for the first column. To
make this happen TreeTable prevents the default on dragenter event.
 - This keyboard handling might not fit for all aggregation reordering
e.g. for tree structures therefore it is now behind a "keyboardHandling"
property with default value "false" in the DragDropBase.
 - Reordering detection of the central DnD concept is explained in the
API documentation of this property as it is only for
dropPosition="Between" setting.
 - mdc/DragDropConfig object is aligned with this new property
 - Mouse/pointer originated drag operation can be started from an inner
control e.g. from the cell content although only the row is draggable
but keyboard handling version of DnD is not getting triggered from inner
DOM nodes unless that DOM node is defined as handle via
data-sap-ui-draghandle. This means m.Table or ui.Table can mark their
cells (<td> elements) as a drag handle via "data-sap-ui-draghandle"
attribute to allow Ctrl+ArrowKey to start dragging while the focus is on
the cell.

Jira: CPOUIFTEAMB-2335

Change-Id: I06f8d3b84518519abb95a41b7cbf186bc3f3048c