[FIX] sap.m.IllustratedMessage: Prevent infinite resize
Problem: In rare cases the IllustratedMessage can loop into infinite
resize.

Root Cause: Due to the scrollbar shown and being forced to go to the
smaller breakpoint. Smaller breakpoint detects enough space (now with no
scrollbar) to go to the bigger breakpoint and this is being looped
endlessly.

Solution: We stash the last known width and height for given size and
if we detect them being the same as the new ones for the size, we don't
update the media (don't do the actual resize).

SNOW: DINC0104395
Change-Id: Ifc8ab7620baab1c7ef187a8f4b4b337658291121