[FIX] sap.m.Dialog: The toolbar spacer is no longer hidden on phone
Solving the following issue:
Any ToolbarSpacer inside the dialog footer
was hidden when dialog opened on a phone.
This is causing the issue that footer items can be misaligned for custom footers.

Solution:
The spacers are now visible on phone.
Hiding of the spacer was needed for sap_belize theme, where buttons are stretched.
Same is needed for sap_bluecrystal, which is already separate from base theme and is not affected by this change.
It is not needed for the rest of the themes.

SNOW: DINC0110533
Change-Id: I84ac5cb9422377e160ce17a71303b3a74b888f94