[INTERNAL] dnd: allow preventDefault for dragover event
 - DropInfo configuration now allows preventDefault for the dragover
event. This can be useful to disable dropping on a specific location
e.g. between two concrete table rows.
 - During dragging over an element a dragenter event might occur at the
edge of two DOM nodes and in this case calculated relative drop position
might be a useful information on whether or not to prevent the default
of dragenter event. For that reason dragEnter event now provides
"dropPosition" as an event parameter
 - mdc/DragDropConfig configuration is enhanced to honor this new core
dnd concepts.

Jira: CPOUIFTEAMB-2356

Change-Id: I7a2b48cfa81ef4c6265090c7a76d549b089cf18a