[INTERNAL] sap.f.AvatarGroup: Provide tooltip for group mode
Problem: AvatarGroup does not implement showing tooltip functionality.

Solution: "tooltip" aggregation is inherited, however, there is missing
implementation in the Renderer of the AvatarGroup, that's why the
tooltip is not shown.
Logic for showing a tooltip (if set) on the AvatarGroup, if it is in
"Group" mode only.

JIRA: BGSOFUIPIRIN-6531
Change-Id: I2c7296024386b7b6b6491922dd1c46ea8f9e2d4a