[INTERNAL] sap.ui.core.qunit.odata.v2.ODataModel.integration.qunit: stabilize test "Messages: Remove the control ID..."
http://localhost:8080/resources/sap/ui/test/starter/Test.qunit.html?testsuite=test-resources%2Fsap%2Fui%2Fcore%2Fqunit%2Fodata%2Fv2%2Ftestsuite.odatav2.qunit&test=ODataModel.integration&filter=Messages%3A%20Remove

Failing test was detected within "Test Stability Report".
Not reproducible locally. We think it is a timing issue
and that we simply have to wait some time before checking the
controlIds for the message in the following .then block

PS1: wait until the control value state changed
PS2: commit message adjusted

Change-Id: I1f3d60689f9f71c97baf907eac9c5181572c265a