[FIX] DemoKit: ScrollToElement in ApiiRef master-side corrected
Problem:
Failure to scroll to an element from the tree on the Master-side
of the ApiRef section. This happened because the responsible code
is outdated and needs adaptation with the latest page elements.
In particular:
1) the checks for the need to scroll assume that the tree take the
entire height of the document, while now the master tree takes
a smaller height within a limited container
2) the function attempts to scroll by refering to a non-existing
element (that was deleted in a recent refactoring)

Solution:
-- Corrected the checks for the need to scroll by specifying the real
tree container
-- scrolling is now done via the scroll delegate of the real tree
container
-- replaced the setTimeout [prior to scrolling] with an event listener

Change-Id: I436ba50dc21fed6be376b88200ddf77b234fd06b
SNOW: DINC0134460