[INTERNAL] introduction of sap/ui/fl/.../SACIntegrationUpdateVariant
SAC uses variants within the application as a cache for the queries related to the corresponding filters.
Due to performance optimization, the queries have changed and this API offers the functionality to update the variants without a user interaction on the fly.
This is needed, because the calculation is done on client side and will be done on a variant instance basis:
Each variant will be updates if the user can update the variant and the variant itself is active.

Change-Id: I234859cb1e13caecaec33d09acd7a76fbb8a5c19
JIRA: CLOUDIOI-1175
BCP: n/a
CR-Id: 002075125800001845132024