[FEATURE] sap.m.PlanningCalendar: new noData aggregation
New "noData" aggregation added to PlanningCalendar where an
IllustratedMessage or text can be provided to be displayed
when the PlanningCalendar has no data/rows.

This aggregation is forwarded to the internal sap.m.Table aggregation
with the same name.

Note: When this aggregation is used with the "noDataText" property, this
aggregation will take priority and the text in noDataText will not be
displayed.

JIRA: BGSOFUIBALKAN-8068
Change-Id: Ibc11f307331d1741ecec45b3714f940e3e16b5a6