[INTERNAL] sap/ui/dom/jquery/control: adjust deprecation comments
The jQuery#control API is deprecated since 1.106, but the deprecation
did not lead to the removal of the function in the next major version
as the transformation can't handle the special structure and naming
of jQuery plugins.

By adding a file deprecation and standardizing the JSDoc name, this has
been fixed.

Change-Id: Iec4b7030154ae9367f54f8d17edbf6c0986b84ad