[FIX] sap.m.SinglePlanningCalendarView: update JSDocs
Issue: The method calculateStartDate in SinglePlanningCalendarView is
abstract and doesn't specify the return type in the JS Doc, which causes
errors in TypeScript.

Solution: Update the method's documentation with the expected return
type. Also update the parameter and return types in the other views to
be more specific. Also update return types for other methods and change
logging to throw error since methods are abstract.

Fixes: https://github.com/SAP/openui5/issues/4045

SNOW: DINC0144944
Change-Id: I78d8271380129caf398245df0716aac7d8a7aa73