[FIX] Tables: Announce required state of columns with 'custom' label controls
If a column has a control other than sap.m.Label, the required state was not announced correctly.
This was the case for the MDCTable, which uses mdc.table.ColumnHeaderLabel instead of m.Label.
Adjusted the logic in both m.Table and ui.Table to announce the state, if the required property
and getAccessibilityInfo is present.

SNOW: DINC0115442
CR-Id: 002075125900002671692024
Change-Id: I314475d6cf41eedbad72d6e160f44c10dc0c3018
(cherry picked from commit ee367af2c5ae4a06620c19a7ff1fcc1d3060dde6)