[FIX] sap.m.SinglePlanningCalendarMonthGrid: remove unnecessary scrolling
Issue: Focusing on a day in SinglePlanningCalendar in a layout will
cause the page to scroll to the entire calendar, even if the focused
item is fully visible on the page. This would happen because
getFocusedDomRef() would return the entire grid not the day cell that
was focused.

The issue was introduced in change #6029833
Snippix #198954

Solution: Return the item navigation focused item instead.

SNOW: DINC0075486
Change-Id: Ie92402a85c314516284eb5557b1bb80f0b1528c3