[FEATURE] sap.m.Carousel: the page indicator can be placed over the content of the Carousel
There is newly created enumeration sap.m.CarouselPageIndicatorPlacementType that replaces
the old sap.m.PlacementType enumeration, which was originally created for the Popover control.
The sap.m.PlacementType contains a lot of values, that are not usable for the Carousel.
Also, with this BLI we add two more values for the pageIndicatorPlacement property,
which are not proper for usage in the Popover - OverContentTop and OverContentBottom.

JIRA: BGSOFUIRODOPI-3213
Change-Id: I4cd3906d1bf5bda79510998cf4e1f7f1e889f24f