[FIX] sap.ui.model.type.(Currency|Unit): use amount scale default 0 instead of currency's or unit's decimal places
The number of decimal places to use for formatting the numeric part of a  unit or currency is in the OData context the minimum of (a) the scale constraint defined for the sap.ui.model.odata.type.Decimal of the numeric part and (b) the number of decimal places for the currency/unit being used. With this commit, the case the the scale from (a) does not exist and thus defaults to 0 is fixed: Before, the scale was ignored and the currency/unit decimals from (b) were always used, possibly leading to a large number of decimals in the formatted value; with the fix, a non-existing scale results in 0 decimal places as expected.

PS1: downport
PS2: provide more details in commit message

Change-Id: Ic1cdc6556e34a98fddc831f9765320ce5526178e
CR-Id: 002075125800002488082024
SNOW: DINC0152691
(cherry picked from commit 86ef554a744acd04b60779392c8091619d44bcb6)