[INTERNAL] qunit-coverage-istanbul: avoid probing of `sap` or `sap.ui`
When `sap` or `sap.ui` are probed separately, the future proof tooling
does not detect that they are part of a usage of `sap.ui.loader`, which
is an allowed global name.

To prevent false positives, the probing is rewritten using `globalThis`
and optional chaining. Function-wise, this is equivalent to the previous
check.

Change-Id: I5f088201e8271661d2111ca2435892bea25f962c