[INTERNAL] testsuite.rule.qunit.js: don't use debug mode for model rules
When executed in the local development setup, the debug mode causes
additional warning logs because the local setup does not produce
*-dbg.js files.

The model/* support rule tests monitor the log and expect no warnings
or errors to occur. To not break these expectations, the debug mode is
switched off for those rule tests.

Change-Id: I01eac689b95a1d576086e156924227b652df4e01