[INTERNAL][FIX] sap.ui.fl Fix issue of loading versions with cacheKey
When the active version of an application has no change. A special
cacheKey <NO CHANGES> can be carried along to connectors even if the
request itself is called in key user mode with a specific version. It is
an issue because LRepConnector will mistake the wrong indicator and skip
sending request

Change-Id: I797a1d3e965a3e9ec1abcba6c536c858eaedf9b4
SNOW: DINC0154902