[FIX] Demokit: IndexEntry controller definition corrected
Problem:
The controller of the IndexEntry block coinsides with the block
itself (probably as a perf optimization, to reduce the number of
created objects). This, however, contradicts with best practice
and will not be allowed in 2.0.

Solution:
Define the controller separately.
--The full name of the new controller includes "controller"
twice on a purpose, as already done for all controller
definitions for blocks in the repo. A research for the
possibility to support the shorter "IndexEntry.controller.js"
name can be done in the future.
--An alternative solution to reuse the controller of the entire
view (instead of defining a new controller) was also tried but
seems not to be supported for now. Further invesigation can be done
in the future. Merging this version of the change with priority
to clear the mentioned incompatibility with 2.0.

Change-Id: I5f3f5ee0261609acd039cf8549f252974bb7d967