[INTERNAL][FIX] sap.ui.fl: Fix timing issue during initial variant switch
In certain scenarios, e.g. when having multiple variant management
controls with many changes on the same page, it can happen that the
order in which switchVariant and registerToModel events are fired gets
mixed up.
This can lead to timing issues, where both methods block each other
because they are part of the same promise chain but wait for the other
promise to finish first before resolving themselves.

To fix this, make sure that the initial waitForInitialVariantChanges
ignores variant switches, since it should always be independent of them
anyways.

SNOW: CS20240007580162
Change-Id: I2e75134a2e671f7f9a89eee93f79a0cf84c745c5
CR-Id: 002075125900001479592024