[FIX] Field: parsing nested brackets
If a user enters text with nested brackets into a Field that should show key and description it is hard to determine what part ist the key and what part is the description.
We assume that the key do not contain any brackets, so we assume that every bracket is part of the description. This might be wrong for special cases, but without knowing the data it is not possible to support every possible case.

Change-Id: I6c1d998fedbb5b4fe399b2c5b7befc6829790de8
SNOW: DINC0128587