[INTERNAL][FIX]sap.ui.fl: Set standard variant as default variant for the compVariant
It was not possible to set the standard variant as default variant since
the standard variant is not in the list of variant in the CompVariantsMap.

The getDefaultVariantId() method was adjusted to include the standard
variant to the check.

CR-Id: 002075125900001605302024
SNOW: DINC0163427
Change-Id: I092bcaa2ae4a8aee5beaede5ca9ed14452d905f4
(cherry picked from commit 2bee8d89f3a82513a9fd82b780eb5b703438bdb5)