[FIX] sap.uxap.ObjectPageLayout: Inconsistent header-collapsing corrected
Problem:
Automatic collapsing of the header upon switching tabs
(without the user explicitly requesting to snap the
header and without the user requesting to scroll to
a lower subsection inside the newly opened tab).

This happens in the cases when in the previous tab the
user already:
1) scrolled down
and
2) expanded the header with a click while preserving
the current scroll position.
=> both are the case:
1) anchorBar is snapped in the top non-scrollable area
2) the header is expanded in the top non-scrollable area

Solution:
Corrected the logic in scrollToSection to ensure no
redundant snap with scroll when the content does not
require it

Change-Id: If89365924c5bb2835c7a9b6a223c6d23d6edc42f
SNOW: DINC0169002