[INTERNAL][FIX] MDCTable: Adjust rowPress event for m.Table
If a "rowPress" listener is added or removed from the table, the ResponsiveTableType does not update the "type" proeproperty of the ColumnListItemTemplate.

To solve this, prepareRowPress and cleanupRowPress have been added, which will add/remove the necessary itemPress event.

If the row action is "Navigation", but it is not visible and there is still a rowPress event, the row should be Active instead of Inactive. This has been adjusted as well.

Additionally, instead of overwriting the Template formatter, a copy of the bindingInfo is being made, which leaves the template untouched for future use.

SNOW: DINC0167928
Change-Id: I01ee483df168acc9beeef93c135eb97040476fc6