[INTERNAL][FIX] sap.ui.fl: fix standard variant handling after creating a new variant
Initially, when there are no variants available, the standard variant is
created by the VariantModel. After a variant is created, the standard variant
would be created by the FlexState (InitialPrepareFunctions).
If a new variant is created, then without reload the app is navigated away from
and to again, the Initial preparation does not run again, and no one would create the fake variant.

To overcome this, the VariantModel moves the faked standard variant to the FlexState upon destroy,
if a variant or a UIChange is created in this session.

SNOW: DINC0071689
Change-Id: Id6734d167b7c2dfb2444fbc675e0b950474bb7a0