[INTERNAL][FIX] sap.ui.fl: Set support user for newly created variants
When updating the support user of a newly created variant based on the
response that was returned from the save request, only the current
VariantModel state was mutated.
Make sure to actually update the underlying flex object instead and then
refresh the variants map.

The author of the variant is calculated in the variant flex object
itself, based on the user id or layer.
Thus in case of perso changes it is not affected by the support user
at all and doesn't need to be updated.

SNOW: DINC0135482
Change-Id: I496d4e34ec22b7c3e790814fbc733b5d515a8d6c