[FIX] sap.ui.unified.Currency: the currency is not truncating now
There was a fix width of the currency and in case the text didn't fit it was truncated. Now the width is set to min-width so the larger currency texts and the bigger text spacing can fit and not to be truncated. This way it is readable and fulfills the ACC standard.

BCP: 2370011691

Change-Id: I320e0254d98438fe5743363f818701dd7267ecf9