[INTERNAL][FIX] sap.ui.rta: fix exit of RuntimeAuthoring with draft available
Workaround for 1.121, newer releases work differently and the issue does not occur.
The save has to happen before the version and maxLayer information are cleared from the FlexInfoSession (triggered by
ReloadManager.checkReloadOnExit), otherwise the FlexState.update does not work correctly. The second save afterwards
is necessary to invalidate the cache.

SNOW: CS20240007652197
CR-Id: 002075125900001708032024
Change-Id: Ia560ddab7825e91c9d9f2edec46606cf82e4339f