[FIX] Demo Kit: Fix API reference search result list item navigation
This commit resolves an issue where certain items were not displayed in
the API reference search result list. The problem was observed when a
user searched for items like "sap/ui/test". The root cause was a filter
condition that incorrectly excluded items with slashes in their names.
This led to items such as "module:sap/ui/test/OpaBuilder" not being
displayed in the search results.

SNOW: n/a
Change-Id: I54c8b15b9ebabfaeb46909ec76ae1b9d2a5c810c