[FIX] FormContainer: invalid role rendered in ColumnLayout
As current JAWS versions announce role="region" the Role structure of the Form is redesigned.
The Form itself is rendered using role "region", the FormContainers are rendered with role "form". (As the FormContainers will contain labels and fields.)
Only of the Form contains only one FormContainer without own Title, Toolbar or AriaLabel, the Form renders role="form". In this case the FormContainer renders no role.
For older layout FormContainer without Title, Toolbar or AriaLabel the FormContainers can render no role at all, here the behaviour is unchanged.

A workaround for an old JAWS18 issue is removed as Labels are not read twice any longer if no title exist.

Change-Id: Ib332c2a092ab32ebc6c715675021c39dd4039d23
SNOW: DINC0134224