[INTERNAL] m.Table: Accessibility improvements
 - Wrong announcement of column count: Although aria-colcount is set
where the role="grid" attribute is set it seems like this property is
not honored by screen readers. For performance reasons, the table
renders the highlight column for every row since setting the highlight
property of a single row needs to change the complete table structure.
This empty highlight cell was not rendered for group headers and no-data
row. Therefore colspan count is bigger than columns aggregation length
for those rows but this causes a wrong announcement. This change renders
an empty highlight and navigated cells so that colspan is not bigger
than columns aggregation length.

 - New PageUp behavior: Table column header row can be sticky via API.
PageUp key was moving focus to the header row when PageUp key is pressed
and if the focus is on a row index smaller than page size. When the
header row is sticky, moving focus to the header row via PageUp does not
scroll since the header row is in the view port. Therefore the first
rows of the table was not accessible. In this change the PageUp moves
the focus first to the data row and if the focus is on the data row than
moves it to the header row.

 - The tooltip of the ListItemBase Delete and Detail button now gets its message
bundle key from the sap.m library. This helps subclasses not to redefine
the bundle key if they belong to another library.

SNOW: DINC0073056
SNOW: DINC0132104

CR-Id: 002075125800003303912024
Change-Id: Ic2aeb53f3e836517ddb0ee7bd1186722007c648c