[FIX] sap.ui.integration.widgets.Card: All elements of a List card item are read out
Problem: We were overwriting a functionality coming form the ListBaseItem, that was generating the output of a list item by creating it in a invisible text and adding it in with an aria-labelledby refrence to each item on focus. By removing this funciotnality we were dependent on what the screen reader reads from the HTML. Highlight text is only a visual representation in HTML, so the screen reader does not read it out.

Solution: Return the functionality of the ListItemBase by expanding the getContentAnnouncment, with all available content in list item of the card. Not depending on the screen reader and what it will read, as this can defer by browser and screen reader.

SNOW: DINC0147227
SNOW: DINC0127299
Change-Id: I6118a2b02798b5aaa102c045106ca249330c6fd7