[INTERNAL][FIX] sap.ui.fl: Always keep refresh parameter in iframe URL
When an iframe url contains multiple bindings in its hash they might be
resolved one by one, resulting in multiple setter calls with hash-only
changes.
On the first call, a frame refresh parameter is added, however it iself
seems to be a change of the search query when comparing the old url and
the new url from the second call, thus it is not added again.
As a result the parameter gets removed on every second call which seems
to lead to missing refreshes in the iframe.

SNOW: CS20240007495016
CR-Id: 002075125900001851722024
Change-Id: I4b77ea5f1f03aa60029b3b81d2100eece14743ba