[INTERNAL][FIX] CellSelector: Range limit should not influence selection
When a range limit is set, cells after that range limit (e.g. row 250 with limit 200) were not able to be selected.

Adjusted the bounds normalization to only account for limits, if necessary: e.g. column selection/range selection.

Additionally, remove column selection functionality for now, as it will be handled in another BLI.

SNOW: DINC0176937
Change-Id: I2b6c6f60e208600d69afe3d8cc985f236c861b10