[INTERNAL] ValueHelp: Focus visualization and selection
Compared with ComboBox and MultiComboBox the Field, FilterField and MultiValueField behave different regarding focus behaviour.
While typing the visual focus should stay inside the Field, the matching item is shown as selected in the dropdown. On opening the dropdown via keyboard or navigation via the arrow keys moves the visual focus to the dropdown.

Change-Id: I0d52fec2ea03182dcba4475768f084686aa313bc